+++ packages/atomic/src/components/atomic-results-per-page/atomic-results-per-page.tsx
@@ -13,7 +13,8 @@ import {
   InitializableComponent,
   InitializeBindings,
 } from '../../utils/initialization-utils';
-import {Button} from '../common/button';
+import {createRipple} from '../../utils/ripple';
+import {randomID} from '../../utils/utils';
 
 /**
  * The `atomic-results-per-page` component determines how many results to display per page.
@@ -33,6 +34,7 @@ export class AtomicResultsPerPage implements InitializableComponent {
   private resultPerPage!: ResultsPerPage;
   public searchStatus!: SearchStatus;
   private choices!: number[];
+  private readonly radioGroupName = randomID('atomic-results-per-page-');
 
   @State()
   @BindStateToController('resultPerPage')
@@ -86,23 +88,25 @@ export class AtomicResultsPerPage implements InitializableComponent {
     }
   }
 
-  // TODO: use standard keyboard shortcuts KIT-949
   private buildChoice(choice: number) {
     const isSelected = this.resultPerPage.isSetTo(choice);
 
     return (
-      <Button
-        style="outline-neutral"
-        role="radio"
-        ariaChecked={`${isSelected}`}
-        ariaLabel={this.bindings.i18n.t('display-results-per-page', {
-          results: choice,
-        })}
-        onClick={() => this.resultPerPage.set(choice)}
-        class={`btn-page ${isSelected ? 'selected' : ''}`}
-        part={`button${isSelected && ' active-button'}`}
-        text={choice.toLocaleString(this.bindings.i18n.language)}
-      ></Button>
+      <input
+        key={choice}
+        type="radio"
+        aria-label={choice}
+        checked={isSelected}
+        value={choice}
+        onChange={(e) =>
+          (e.currentTarget as HTMLInputElement).checked &&
+          this.resultPerPage.set(choice)
+        }
+        class={`btn-page btn-outline-neutral ${isSelected ? 'selected' : ''}`}
+        name={this.radioGroupName}
+        onMouseDown={(e) => createRipple(e, {color: 'neutral'})}
+        part={`button${isSelected ? ' active-button' : ''}`}
+      />
     );
   }